// src/serverApi/v3/models/import-user-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { ImportUserResponse } from './import-user-response';

/**
 * 
 * @export
 * @interface ImportUserListResponse
 */
export interface ImportUserListResponse {
    /**
     * The items for the current page.
     * @type {Array<ImportUserResponse>}
     * @memberof ImportUserListResponse
     */
    data: Array<ImportUserResponse>;
    /**
     * The total amount of items.
     * @type {number}
     * @memberof ImportUserListResponse
     */
    total: number;
    /**
     * The amount of items skipped from the start.
     * @type {number}
     * @memberof ImportUserListResponse
     */
    skip: number;
    /**
     * The page size of the response.
     * @type {number}
     * @memberof ImportUserListResponse
     */
    limit: number;
}