// src/serverApi/v3/models/user-match-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { UserMatchResponse } from './user-match-response';

/**
 * 
 * @export
 * @interface UserMatchListResponse
 */
export interface UserMatchListResponse {
    /**
     * The items for the current page.
     * @type {Array<UserMatchResponse>}
     * @memberof UserMatchListResponse
     */
    data: Array<UserMatchResponse>;
    /**
     * The total amount of items.
     * @type {number}
     * @memberof UserMatchListResponse
     */
    total: number;
    /**
     * The amount of items skipped from the start.
     * @type {number}
     * @memberof UserMatchListResponse
     */
    skip: number;
    /**
     * The page size of the response.
     * @type {number}
     * @memberof UserMatchListResponse
     */
    limit: number;
}